On branch action2054-haspopup modified:   aria/aria.html as follows:
Removed sentence: "This means that activation renders conditional content."
Reason: Activation implies default action. not all popups, e.g., context menues, open with the default action.

Changed this sentence into a note:
Note that ordinary tooltips are not considered popups in this context.

Made editorial changes to the description of the appearance of a popup to remove unnecessary words.